scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.14.0
pyppeteer==1.0.2
playwright==1.39.0
fastapi==0.104.1
uvicorn==0.23.2
celery==5.3.4
redis==5.0.1
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
pandas==2.1.1
tqdm==4.66.1
lxml==4.9.3
python-linkedin==0.1.5
user-agent==0.1.10
aiohttp==3.8.6
